+++ lib/served/attribute/base.rb
@@ -1,0 +1,9 @@
+module Served
+  module Attribute
+    class Base
+      include Support::Attributable
+      include Support::Serializable
+      include Support::Validatable
+    end
+  end
+end