+++ lib/served/support.rb
@@ -1,0 +1,3 @@
+require_relative 'support/attributable'
+require_relative 'support/serializable'
+require_relative 'support/validatable'