+++ CHANGELOG.md
@@ -5,7 +5,18 @@
 * use OJ for JSON serialization
 * new HTTP client backend functionality, ability to support multiple
   HTTP backends
-* ability to set the resource_name
+
+## 0.1.11
+ * add validation suppport
+ * add serialization support
+ * allow definition of individual attributes
+ * add resource level `headers` option
+ * add resource level `reasource_name` option
+ * add resource level `host` option, `host_config` will be deprecated in 0.2.0
+ * add `Served::Attribute::Base` class
+
+## 0.1.10
+* make `timeout` configurable per resource
 
 ## 0.1.8